+++ src/__init__.py
@@ -1,4 +1,4 @@
 """Fexobooth - Photobooth Software für fexobox"""
 
-__version__ = "2.4.38"
+__version__ = "2.4.39"
 __author__ = "fexon e.K."